+++ README.md
@@ -30,6 +30,7 @@
 
 ### 예약 가능한 시간 조회
 
+- [x] 예약 가능한 시간을 모두 조회할 수 있다.
 - [x] 예약 여부와 함께 예약 가능 시간을 조회할 수 있다.
 
 ## :rocket: 테마
@@ -74,10 +75,6 @@
 - [x] 예약 가능한 시간을 추가할 수 있다.
     - [x] 이미 존재하는 시간의 경우 추가할 수 없다.
     - [x] 유효하지 않은 요청으로 예약 가능 시간을 추가할 수 없다.
-
-### 예약 가능한 시간 조회
-
-- [x] 예약 가능한 시간을 조회할 수 있다.
 
 ### 예약 가능한 시간 삭제
