+++ src/main/java/roomescape/controller/admin/ThemeController.java
@@ -1,48 +1,25 @@
-package roomescape.controller;
+package roomescape.controller.admin;
 
 import jakarta.validation.Valid;
 import java.net.URI;
-import java.time.LocalDate;
-import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.Theme;
 import roomescape.dto.request.ThemeCreationRequest;
 import roomescape.dto.response.ThemeResponse;
 import roomescape.service.ThemeService;
 
-@RestController
-@RequestMapping
+@RestController(value = "AdminThemeController")
 public class ThemeController {
 
     private final ThemeService themeService;
 
     public ThemeController(ThemeService themeService) {
         this.themeService = themeService;
-    }
-
-    @GetMapping("/themes")
-    public List<ThemeResponse> getThemes() {
-        List<Theme> themes = themeService.findAllTheme();
-        return themes.stream()
-                .map(ThemeResponse::new)
-                .toList();
-    }
-
-    @GetMapping("/themes/top")
-    public List<ThemeResponse> getTopThemes() {
-        LocalDate endDate = LocalDate.now();
-        LocalDate startDate = endDate.minusDays(7);
-        List<Theme> themes = themeService.findTopThemes(startDate, endDate, 10);
-        return themes.stream()
-                .map(ThemeResponse::new)
-                .toList();
     }
 
     @PostMapping("/themes")
