+++ src/main/java/roomescape/controller/user/PageController.java
@@ -1,0 +1,18 @@
+package roomescape.controller.user;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class PageController {
+
+    @GetMapping("/reservation")
+    public String getUserReservationPage() {
+        return "reservation";
+    }
+
+    @GetMapping("/login")
+    public String getLoginPage() {
+        return "login";
+    }
+}
