+++ src/main/java/roomescape/controller/user/ReservationTimeController.java
@@ -1,0 +1,42 @@
+package roomescape.controller.user;
+
+import java.time.LocalDate;
+import java.util.List;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.domain.ReservationTime;
+import roomescape.dto.other.TimeWithBookState;
+import roomescape.dto.response.ReservationTimeResponse;
+import roomescape.dto.response.TimeWithBookStateResponse;
+import roomescape.service.ReservationTimeService;
+
+@RestController
+public class ReservationTimeController {
+
+    private final ReservationTimeService reservationTimeService;
+
+    public ReservationTimeController(ReservationTimeService reservationTimeService) {
+        this.reservationTimeService = reservationTimeService;
+    }
+
+    @GetMapping("/times")
+    public List<ReservationTimeResponse> getReservationTimes() {
+        List<ReservationTime> times = reservationTimeService.getAllReservationTime();
+        return times.stream()
+                .map(ReservationTimeResponse::new)
+                .toList();
+    }
+
+    @GetMapping(value = "/times", params = {"date", "themeId"})
+    public List<TimeWithBookStateResponse> getReservationTimesInThemeAndDate(
+            @RequestParam("date") LocalDate date,
+            @RequestParam("themeId") Long themeId
+    ) {
+        List<TimeWithBookState> times = reservationTimeService
+                .getAllReservationTimeWithBookState(date, themeId);
+        return times.stream()
+                .map(TimeWithBookStateResponse::new)
+                .toList();
+    }
+}
